refactor: Mandos renders from the block registry
Extracts the controller config + scope into a useController hook shared by the
manager/remap/settings blocks through a MandosProvider (the scope tab and per-game
remap are cross-block state). Availability of the remap vs settings block is
state-derived from the live daemon kind (InputPlumber vs HHD).
